sms: Mapper fixup

+++ Source/main.c
@@ -86,7 +86,7 @@ static uint8_t sms_memory_read (uint16_t addr)
     /* Cartridge, card, BIOS, expansion slot */
     if (addr >= 0x0000 && addr <= 0xbfff)
     {
-        uint16_t bank_base = mapper_bank[(addr >> 14)] * (16 << 10);
+        uint32_t bank_base = mapper_bank[(addr >> 14)] * ((uint32_t)16 << 10);
         uint16_t offset    = addr & 0x3fff;
         /* if ((memory_control & 0x08) == 0x00) */
             return bios[bank_base + offset];
